+++ lib/clixon/clixon_yang.h
@@ -282,5 +282,7 @@ int        yang_extension_value(yang_stmt *ys, char *name, char *ns, int *exist,
 int        yang_sort_subelements(yang_stmt *ys);
 int        yang_init(clicon_handle h);
 int        yang_single_child_type(yang_stmt *ys, enum rfc_6020 subkeyw);
+void      *yang_action_cb_get(yang_stmt *ys);
+int        yang_action_cb_add(yang_stmt *ys, void *rc);
 
 #endif  /* _CLIXON_YANG_H_ */