+++ lib/src/clixon_yang_parse_lib.c
@@ -85,7 +85,6 @@
 #include "clixon_handle.h"
 #include "clixon_file.h"
 #include "clixon_yang.h"
-#include "clixon_yang_internal.h"
 #include "clixon_hash.h"
 #include "clixon_xml.h"
 #include "clixon_xml_nsctx.h"
@@ -98,6 +97,8 @@
 #include "clixon_yang_type.h"
 #include "clixon_yang_parse.h"
 #include "clixon_yang_cardinality.h"
+#include "clixon_plugin.h"
+#include "clixon_yang_internal.h"
 #include "clixon_yang_parse_lib.h"
 
 /* Size of json read buffer when reading from file*/